@@ -1,7 +1,7 @@
 name: dartbenchmark
 description: A benchmark of dart
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.0.0
+  lints: ^2.0.0
