@@ -1,13 +1,11 @@
 package com.test.io;
 
-import com.test.io.handler.JsonHandler;
-import com.test.io.handler.TextHandler;
 import io.magician.Magician;
-import io.magician.common.event.EventGroup;
+import io.magician.common.config.MagicianConfig;
+import io.magician.network.HttpServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.concurrent.Executors;
 
 public class Start {
 
@@ -16,13 +14,17 @@ public class Start {
     public static void main(String[] args) {
         try {
 
-            EventGroup ioEventGroup = new EventGroup(1, Executors.newCachedThreadPool());
-            EventGroup workerEventGroup = new EventGroup(4, Executors.newCachedThreadPool());
+            MagicianConfig magicianConfig = new MagicianConfig();
+            magicianConfig.setNumberOfPorts(1);
+            magicianConfig.setBossThreads(3);
+            magicianConfig.setWorkThreads(5);
 
-            Magician.createTCPServer(ioEventGroup, workerEventGroup)
-                    .handler("/json", new JsonHandler())
-                    .handler("/plaintext", new TextHandler())
-                    .bind(8080, 10000);
+
+            HttpServer httpServer = Magician.createHttp()
+                    .scan("com.test.io")
+                    .setConfig(magicianConfig);
+
+            httpServer.bind(8080);
 
         } catch (Exception e){
             logger.error("启动服务出现异常", e);
