@@ -1,11 +1,12 @@
 package hello.jpa;
 
-import hello.model.Fortune;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
-@Component
+import hello.model.Fortune;
+
+@Repository
 @Profile("jpa")
 public interface FortuneRepository extends JpaRepository<Fortune, Integer> {
 }
