@@ -1,11 +1,12 @@
 package hello.jpa;
 
-import hello.model.World;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
-@Component
+import hello.model.World;
+
+@Repository
 @Profile("jpa")
 public interface WorldRepository extends JpaRepository<World, Integer> {
 }
