@@ -1,14 +1,14 @@
 package hello.repository;
 
+import java.util.List;
+
 import hello.model.Fortune;
 import hello.model.World;
 
-import java.util.List;
-
 public interface DbRepository {
-    World getWorld(int id);
+	World getWorld(int id);
 
-    World updateWorld(World world, int randomNumber);
+	World updateWorld(World world, int randomNumber);
 
-    List<Fortune> fortunes();
+	List<Fortune> fortunes();
 }
