@@ -1,3 +1,4 @@
+use axum::async_trait;
 use axum::extract::{Extension, FromRequest, RequestParts};
 use axum::http::StatusCode;
 use futures_util::stream::FuturesUnordered;
@@ -5,7 +6,7 @@ use futures_util::TryStreamExt;
 use std::io;
 
 use crate::utils::internal_error;
-use crate::{World};
+use crate::World;
 use mongodb::bson::{doc, RawDocumentBuf};
 use mongodb::Database;
 
@@ -57,8 +58,19 @@ pub async fn find_world_by_id(db: Database, id: i32) -> Result<World, MongoError
         .expect("expected world, found none");
 
     Ok(World {
-        id: raw.get("id").expect("expected to parse world id").expect("could not get world id").as_f64().expect("could not extract world id") as f32,
-        random_number: raw.get("randomNumber").expect("expected to parse world randomNumber").expect("expected to get world randomNumber").as_f64().expect("could not extract world randomNumber") as f32,
+        id: raw
+            .get("id")
+            .expect("expected to parse world id")
+            .expect("could not get world id")
+            .as_f64()
+            .expect("could not extract world id") as f32,
+        random_number: raw
+            .get("randomNumber")
+            .expect("expected to parse world randomNumber")
+            .expect("expected to get world randomNumber")
+            .as_f64()
+            .expect("could not extract world randomNumber")
+            as f32,
     })
 }
 
