@@ -1,3 +1,4 @@
+use axum::async_trait;
 use axum::extract::{Extension, FromRequest, RequestParts};
 use axum::http::StatusCode;
 use std::io;
@@ -70,7 +71,6 @@ pub async fn fetch_world(
         sqlx::query_as_with("SELECT id, randomnumber FROM World WHERE id = $1", args)
             .fetch_one(&mut conn)
             .await
-            .ok()
             .expect("error loading world");
     Ok(world)
 }
@@ -81,7 +81,6 @@ pub async fn fetch_fortunes(
     let fortunes: Vec<Fortune> = sqlx::query_as("SELECT * FROM Fortune")
         .fetch_all(&mut conn)
         .await
-        .ok()
         .expect("error loading Fortunes");
     Ok(fortunes)
 }
