@@ -1,15 +1,17 @@
 use std::io;
 use std::net::{Ipv4Addr, SocketAddr};
 
-use salvo::hyper::server::conn::AddrIncoming;
 use salvo::hyper;
+use salvo::hyper::server::conn::AddrIncoming;
 use tokio::net::{TcpListener, TcpSocket};
 
 pub fn builder() -> hyper::server::Builder<AddrIncoming> {
     let addr = SocketAddr::from((Ipv4Addr::UNSPECIFIED, 8080));
     let listener = reuse_listener(addr).expect("couldn't bind to addr");
     let incoming = AddrIncoming::from_listener(listener).unwrap();
-    hyper::Server::builder(incoming).http1_only(true).tcp_nodelay(true)
+    hyper::Server::builder(incoming)
+        .http1_only(true)
+        .tcp_nodelay(true)
 }
 
 fn reuse_listener(addr: SocketAddr) -> io::Result<TcpListener> {
@@ -19,11 +21,11 @@ fn reuse_listener(addr: SocketAddr) -> io::Result<TcpListener> {
     };
 
     #[cfg(unix)]
-        {
-            if let Err(e) = socket.set_reuseport(true) {
-                eprintln!("error setting SO_REUSEPORT: {}", e);
-            }
+    {
+        if let Err(e) = socket.set_reuseport(true) {
+            eprintln!("error setting SO_REUSEPORT: {}", e);
         }
+    }
 
     socket.set_reuseaddr(true)?;
     socket.bind(addr)?;
